# lib/api-zod/src/generated/types/call.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallOutcome } from "./callOutcome";

export interface Call {
  id: number;
  callerName?: string | null;
  callerPhone?: string | null;
  duration?: number | null;
  outcome?: CallOutcome;
  summary?: string | null;
  transcript?: string | null;
  extractedLeadId?: number | null;
  calledAt: Date;
  createdAt: Date;
}