# lib/api-zod/src/generated/types/updateOnboardingProgressBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOnboardingProgressBody {
  currentStep: number;
  completedStep?: number;
}